Wire goal planning frontend fixes
