+++ apps/frontend/src/adapters/web/index.ts
@@ -106,6 +106,7 @@ export {
   getGoals,
   getRetirementOverview,
   getSaveUpOverview,
+  previewSaveUpOverview,
   refreshAllGoalSummaries,
   refreshGoalSummary,
   saveGoalFunding,
